% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osti.R
\name{osti_record}
\alias{osti_record}
\title{Fetch a single OSTI record by ID}
\usage{
osti_record(id)
}
\arguments{
\item{id}{OSTI record ID}
}
\value{
tibble: single row with record details
}
\description{
Fetch a single OSTI record by ID
}
